+++ C++/graph/ShortestPath.hpp
@@ -7,10 +7,10 @@
 namespace man {
 struct ShortestPath {
 private:
-    const std::vector<int64_t> cost;
+    const std::vector<long long> cost;
     const std::vector<int> src;
 public:
-    ShortestPath(const std::vector<int64_t> &cost, const std::vector<int> &src): cost(cost), src(src){}
+    ShortestPath(const std::vector<long long> &cost, const std::vector<int> &src): cost(cost), src(src){}
     inline bool is_thru(const int i) const noexcept { return src[i] != -1; }
     inline std::vector<int> path(int i) noexcept {
         std::vector<int> ret;
@@ -20,7 +20,7 @@ public:
         std::ranges::reverse(ret);
         return ret;
     }
-    inline std::vector<int64_t> get() const noexcept { return cost; }
+    inline std::vector<long long> get() const noexcept { return cost; }
 };
 }
