+++ C++/math/divisor.hpp
@@ -5,7 +5,7 @@
 namespace man {
 template <std::integral T> inline std::vector<T> div(const T n) noexcept {
     std::vector<T> d;
-    for(int64_t i = 1; i * i <= n; ++i) {
+    for(long long i = 1; i * i <= n; ++i) {
         if(n % i == 0) {
             d.emplace_back(i);
             if(i * i != n) {
