+++ C++/math/primecounter.hpp
@@ -12,24 +12,24 @@ template <class T> constexpr inline T cub(const T x) noexcept { return x * x * x
 namespace man {
 struct p_count {
 private:
-    int64_t sq;
+    long long sq;
     std::vector<int> prime;
-    std::vector<int64_t> prime_sum, primes;
-    constexpr inline int64_t p2(const int64_t x, const int64_t y) noexcept {
+    std::vector<long long> prime_sum, primes;
+    constexpr inline long long p2(const long long x, const long long y) noexcept {
         if(x < 4) {
             return 0;
         }
-        const int64_t a = pi(y), b = pi(kthrooti(x, 2));
+        const long long a = pi(y), b = pi(kthrooti(x, 2));
         if(a >= b) {
             return 0;
         }
-        int64_t sum = (a - 2) * (a + 1) / 2 - (b - 2) * (b + 1) / 2;
+        long long sum = (a - 2) * (a + 1) / 2 - (b - 2) * (b + 1) / 2;
         for(const auto i: std::views::iota(a, b)) {
             sum += pi(x / primes[i]);
         }
         return sum;
     }
-    constexpr inline int64_t phi(const int64_t m, const int64_t n) noexcept {
+    constexpr inline long long phi(const long long m, const long long n) noexcept {
         if(m < 1) {
             return 0;
         }
@@ -43,8 +43,8 @@ private:
             return pi(m) - n + 1;
         }
         if(m <= cub(primes[n - 1]) && m <= sq) {
-            const int64_t sx = pi(kthrooti(m, 2));
-            int64_t ans = pi(m) - (sx + n - 2) * (sx - n + 1) / 2;
+            const long long sx = pi(kthrooti(m, 2));
+            long long ans = pi(m) - (sx + n - 2) * (sx - n + 1) / 2;
             for(const auto i: std::views::iota(n, sx)) {
                 ans += pi(m / primes[i]);
             }
@@ -53,7 +53,7 @@ private:
         return phi(m, n - 1) - phi(m / primes[n - 1], n - 1);
     }
 public:
-    p_count(const int64_t lim): sq(kthrooti(lim, 2)), prime_sum(sq + 1) {
+    p_count(const long long lim): sq(kthrooti(lim, 2)), prime_sum(sq + 1) {
         prime = p_table(sq).SoE;
         for(const auto i: std::views::iota(1) | std::views::take(sq)) {
             prime_sum[i] = prime_sum[i - 1] + prime[i];
@@ -65,12 +65,12 @@ public:
             }
         }
     }
-    inline int64_t pi(const int64_t n) noexcept {
+    inline long long pi(const long long n) noexcept {
         if(n <= sq) {
             return prime_sum[n];
         }
-        const int64_t m = kthrooti(n, 3);
-        const int64_t a = pi(m);
+        const long long m = kthrooti(n, 3);
+        const long long a = pi(m);
         return phi(n, a) + a - 1 - p2(n, m);
     }
 };
