+++ C++/math/primefactor.hpp
@@ -5,7 +5,7 @@
 namespace man {
 template <std::integral T> inline std::vector<std::pair<T, int>> prmfct(T n) noexcept {
     std::vector<std::pair<T, int>> ret;
-    for(int64_t i = 2; i * i <= n; ++i) {
+    for(long long i = 2; i * i <= n; ++i) {
         if(n % i != 0) {
             continue;
         }
