+++ C++/math/primefactortable.hpp
@@ -5,12 +5,12 @@
 #include <numeric>
 namespace man {
 struct p_fact {
-    std::vector<int64_t> spf;
+    std::vector<long long> spf;
     p_fact(const int n): spf(n + 1){
         std::iota(spf.begin(), spf.end(), 0);
-        for(int64_t i = 2; i * i <= n; ++i) {
+        for(long long i = 2; i * i <= n; ++i) {
             if(spf[i] == i) {
-                for(int64_t j = i * i; j <= n; j += i) {
+                for(long long j = i * i; j <= n; j += i) {
                     if(spf[j] == j) {
                         spf[j] = i;
                     }
